Some more progress made but some issues have arisen
I have made lvl calculations automatic.

However due to dependency of some tests on other tests (this is
probably something we really don't want since it is a bad test-suite
design pattern) we have to modify the dune rules to output to the log
rather than the target since there are now multiple targets.

There is unfortunately a problem with this setup as there is a file
called "test-suite/modules/pseudo_circular_with.v" which as the name
suggests induces circular deps. :/

This makes dune have a headache so I don't know what to do here
really. Perhaps we should forbid dependencies between tests anyway,
but then the module tests will need some refactoring. The alternative
is to keep it as is but remove these weird module tests.