[coqmod] Added new binary coqmod
coqmod is a simple lexer for .v files which outputs parsed tokens that
can be interpreted as dependencies by another program. It is intended
as a replacement for coqdep for use in programs such as Dune.

Unlike coqdep, it does not scan directories making it lightweight in
comparison.

It has 3 --format modes:
- csexp - canonical serial expresion (default)
- sexp - pretty printed serial expression
- read - human readable output

See the README.md for more information.

Signed-off-by: Ali Caglayan <alizter@gmail.com>